Change cargo version check to use shell=True
I'll admit I'm not very handy with Windows shells in general, but I would like to make a package I help maintain installable for all users, Windows included. I have confirmed that `cargo.exe` is in fact present as expected, it's just that this subshell doesn't seem to find it.

Using `shell=True` appears to work on Windows.